Added error checking and comments
comments are for error checking and can be found at the bottom of the server code. I tried to comment everwhere I made changes although I missed a lot realizing that git will handle that.
To do now is really just sending the file keeping in mind that datagram sockets only send one package at a time when you call send and only receive one packet at a time when you call that too.
So it needs to be in a loop of 5 because that is our shifting window size.