updated notes
I added some notes from lab on how to implement the sliding window because the way I was attempting it was completely wrong (I was using the position number in the window, not the actual packet number). So, I'm having to start over. If you have any questions on the notes let me know.